requirements.txt
# Core FastAPI and web framework
fastapi==0.108.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6

# Pydantic for data validation
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP requests
requests==2.31.0
httpx==0.26.0

# Redis for session management (using redis-py instead of redis[hiredis] for compatibility)
redis==5.0.1

# OpenAI for Whisper (Speech-to-Text)
openai==1.6.1

# Google Cloud services
google-cloud-translate==3.13.0
google-cloud-texttospeech==2.16.4
googlemaps==4.10.0

# AWS SDK for Bedrock
boto3==1.34.34
botocore==1.34.34

# Audio processing (simplified for compatibility)
numpy==1.26.3

# Environment management
python-dotenv==1.0.0

# Date and time utilities
python-dateutil==2.8.2

# JSON handling
ujson==5.9.0

# Security
cryptography==42.0.0

# Testing (development only)
pytest==7.4.4
pytest-asyncio==0.23.3

# Code quality (development only)
black==23.12.1
isort==5.13.2